@@ -1,14 +1,18 @@
-[tool.poetry]
+[project]
 name = "graphemy"
 version = "1.0.4"
 description = "A Python library for integrating SQLModel and Strawberry, providing a seamless GraphQL integration with FastAPI and advanced features for database interactions."
-authors = ["Matheus Doreto <matheusdoreto.md@gmail.com>"]
+authors = [
+    {name = "Matheus Doreto", email = "matheusdoreto.md@gmail.com"},
+]
+license = {text = "MIT"}
+requires-python = "<4.0,>=3.12"
+dependencies = [
+    "sqlmodel<1.0.0,>=0.0.22",
+    "strawberry-graphql[debug-server]<1.0.0,>=0.240.0",
+    "fastapi<1.0.0,>=0.114.0",
+]
 readme = "README.md"
-packages = [{include = "graphemy"}]
-homepage = "https://github.com/MDoreto/graphemy"
-documentation = "https://graphemy.readthedocs.io/en/latest/"
-repository = "https://github.com/MDoreto/graphemy"
-license = "MIT"
 classifiers = [
     "Development Status :: 5 - Production/Stable",
     "Topic :: Database",
