@@ -1,4 +1,4 @@
-import os
+from pathlib import Path
 
 from fastapi import FastAPI
 from sqlmodel import create_engine
@@ -6,7 +6,7 @@ from sqlmodel.pool import StaticPool
 
 from graphemy import Graphemy, GraphemyRouter, import_files
 
-import_files(os.path.dirname(__file__))
+import_files(Path(__file__).parent)
 
 engine = create_engine(
     "sqlite://",
@@ -18,6 +18,8 @@ Graphemy.metadata.create_all(engine)
 
 app = FastAPI()
 router = GraphemyRouter(
-    engine=engine, enable_put_mutations=True, enable_delete_mutations=True
+    engine=engine,
+    enable_put_mutations=True,
+    enable_delete_mutations=True,
 )
 app.include_router(router, prefix="/graphql")
