@@ -18,8 +18,8 @@ class Owner(Graphemy, table=True):
     privates: list["Private"] = Dl(source="id", target="owner_id")
     keys: list["Key"] = Dl(source="id", target="owner_id")
 
-    async def permission_getter(context: dict, request_type: str) -> bool:
-        return "Owner" in context["user"]["classes"]
+    async def permission_getter(self: dict, _request_type: str) -> bool:
+        return "Owner" in self["user"]["classes"]
 
 
 class Private(Graphemy, table=True):
