@@ -7,6 +7,8 @@ class Teacher(Graphemy, table=True):
     id: int | None = Field(primary_key=True, default=None)
     name: str
     courses: list["Course"] = Dl(source="id", target="teacher_id")
+    school_id: int
+    school: "School" = Dl(source="school_id", target="id")
 
 
 class Course(Graphemy, table=True):
@@ -19,9 +21,11 @@ class Course(Graphemy, table=True):
 
 class Student(Graphemy, table=True):
     id: int | None = Field(primary_key=True, default=None)
-    name: str
+    name: str | None
     birth_date: date
     courses: list["StudentCourse"] = Dl(source="id", target="student_id")
+    school_id: int
+    school: "School" = Dl(source="school_id", target="id")
 
 
 class StudentCourse(Graphemy, table=True):
@@ -30,7 +34,8 @@ class StudentCourse(Graphemy, table=True):
     student: "Student" = Dl(source="student_id", target="id")
     course: "Course" = Dl(source="course_id", target="id")
     grader: list["Grade"] = Dl(
-        source=["student_id", "course_id"], target=["student_id", "course_id"]
+        source=["student_id", "course_id"],
+        target=["student_id", "course_id"],
     )
 
 
@@ -40,5 +45,13 @@ class Grade(Graphemy, table=True):
     grade: float
     semester: int = Field(primary_key=True)
     student_course: "StudentCourse" = Dl(
-        source=["student_id", "course_id"], target=["student_id", "course_id"]
+        source=["student_id", "course_id"],
+        target=["student_id", "course_id"],
     )
+
+
+class School(Graphemy, table=True):
+    id: int | None = Field(primary_key=True, default=None)
+    name: str
+    students: list["Student"] = Dl(source="id", target="school_id")
+    teachers: list["Teacher"] = Dl(source="id", target="school_id")
