@@ -3,12 +3,12 @@ def test_date_year(client_data):
         "/graphql",
         json={
             "query": """query MyQuery {
-                            students (filters: {birthDate: {year: 1999}}){
+                            students (where: {birthDate: {gte: "1999-01-01", lt: "2000-01-01"}}) {
                                 id
                                 name
                                 birthDate
                                 }
-                            }"""
+                            }""",
         },
     )
     assert response.status_code == 200
@@ -17,8 +17,8 @@ def test_date_year(client_data):
             "students": [
                 {"id": 1, "name": "Some Name", "birthDate": "1999-09-16"},
                 {"id": 2, "name": "Other Name", "birthDate": "1999-07-24"},
-            ]
-        }
+            ],
+        },
     }
 
 
@@ -27,12 +27,12 @@ def test_date_range(client_data):
         "/graphql",
         json={
             "query": """query MyQuery {
-                            students (filters: {birthDate: {range: ["1998-04-01","1999-08-25"]}}) {
+                            students (where: {birthDate: {gt: "1998-04-01", lte: "1999-08-25"}}) {
                                 id
                                 name
                                 birthDate
                                 }
-                            }"""
+                            }""",
         },
     )
     assert response.status_code == 200
@@ -41,8 +41,8 @@ def test_date_range(client_data):
             "students": [
                 {"id": 2, "name": "Other Name", "birthDate": "1999-07-24"},
                 {"id": 3, "name": "Another Name", "birthDate": "1998-05-12"},
-            ]
-        }
+            ],
+        },
     }
 
 
@@ -51,12 +51,12 @@ def test_date_items(client_data):
         "/graphql",
         json={
             "query": """query MyQuery {
-                            students (filters: {birthDate: {items: "1999-09-16"}}){
+                            students (where: {birthDate: {in: "1999-09-16"}}){
                                 id
                                 name
                                 birthDate
                                 }
-                            }"""
+                            }""",
         },
     )
     assert response.status_code == 200
@@ -64,8 +64,8 @@ def test_date_items(client_data):
         "data": {
             "students": [
                 {"id": 1, "name": "Some Name", "birthDate": "1999-09-16"},
-            ]
-        }
+            ],
+        },
     }
 
 
@@ -79,12 +79,12 @@ def test_date_nested_year(client_data):
     name
     teacherId
     students {
-      student(filters: {birthDate: {year: 1998}}) {
+      student(where: {birthDate: {gte: "1998-01-01", lt: "1999-01-01"}}) {
         name
       }
     }
   }
-}"""
+}""",
         },
     )
     assert response.status_code == 200
@@ -107,8 +107,8 @@ def test_date_nested_year(client_data):
                     "teacherId": 1,
                     "students": [{"student": None}, {"student": None}],
                 },
-            ]
-        }
+            ],
+        },
     }
 
 
@@ -122,12 +122,12 @@ def test_date_nested_range(client_data):
     name
     teacherId
     students {
-      student(filters: {birthDate: {range: ["1999-08-01","1999-10-25"]}}) {
+      student(where: {birthDate: {gt:"1999-08-01", lte:"1999-10-25"}}) {
         name
       }
     }
   }
-}"""
+}""",
         },
     )
     assert response.status_code == 200
@@ -153,8 +153,8 @@ def test_date_nested_range(client_data):
                         {"student": None},
                     ],
                 },
-            ]
-        }
+            ],
+        },
     }
 
 
@@ -168,12 +168,12 @@ def test_date_nested_items(client_data):
     name
     teacherId
     students {
-      student(filters: {birthDate: {items: ["1999-07-24"]}}) {
+      student(where: {birthDate: {in: ["1999-07-24"]}}) {
         name
       }
     }
   }
-}"""
+}""",
         },
     )
     assert response.status_code == 200
@@ -199,6 +199,6 @@ def test_date_nested_items(client_data):
                         {"student": {"name": "Other Name"}},
                     ],
                 },
-            ]
-        }
+            ],
+        },
     }
