@@ -8,7 +8,7 @@ def test_query(client_data):
                                 name
                                 birthDate
                                 }
-                            }"""
+                            }""",
         },
     )
     assert response.status_code == 200
@@ -18,8 +18,8 @@ def test_query(client_data):
                 {"id": 1, "name": "Some Name", "birthDate": "1999-09-16"},
                 {"id": 2, "name": "Other Name", "birthDate": "1999-07-24"},
                 {"id": 3, "name": "Another Name", "birthDate": "1998-05-12"},
-            ]
-        }
+            ],
+        },
     }
 
 
@@ -28,12 +28,12 @@ def test_query_filter(client_data):
         "/graphql",
         json={
             "query": """query MyQuery {
-                            students (filters: {id: 1}){
+                            students (where: {id: {in:1}}){
                                 id
                                 name
                                 birthDate
                                 }
-                            }"""
+                            }""",
         },
     )
     assert response.status_code == 200
@@ -41,8 +41,8 @@ def test_query_filter(client_data):
         "data": {
             "students": [
                 {"id": 1, "name": "Some Name", "birthDate": "1999-09-16"},
-            ]
-        }
+            ],
+        },
     }
 
 
@@ -59,7 +59,7 @@ def test_relationship(client_data):
                                     name
                                     }
                                 }
-                            }"""
+                            }""",
         },
     )
     assert response.status_code == 200
@@ -73,9 +73,9 @@ def test_relationship(client_data):
                         {"id": 1, "name": "Mathematics"},
                         {"id": 2, "name": "Physics"},
                     ],
-                }
-            ]
-        }
+                },
+            ],
+        },
     }
 
 
@@ -87,12 +87,12 @@ def test_relationship_filter(client_data):
                             teachers {
                                 id
                                 name
-                                courses (filters: {id: 1}){
+                                courses (where: {id: {in:1 } }){
                                     id
                                     name
                                     }
                                 }
-                            }"""
+                            }""",
         },
     )
     assert response.status_code == 200
@@ -105,9 +105,9 @@ def test_relationship_filter(client_data):
                     "courses": [
                         {"id": 1, "name": "Mathematics"},
                     ],
-                }
-            ]
-        }
+                },
+            ],
+        },
     }
 
 
@@ -126,7 +126,7 @@ def test_relationship_one(client_data):
     }
   }
 }
-                            """
+                            """,
         },
     )
     assert response.status_code == 200
@@ -145,8 +145,8 @@ def test_relationship_one(client_data):
                     "teacherId": 1,
                     "teacher": {"id": 1, "name": "Some Teacher"},
                 },
-            ]
-        }
+            ],
+        },
     }
 
 
@@ -159,13 +159,13 @@ def test_relationship_one_filter(client_data):
     id
     name
     teacherId
-    teacher (filters: {id: 2}){
+    teacher (where: {id: { in: 2 } }){
       id
       name
     }
   }
 }
-                            """
+                            """,
         },
     )
     assert response.status_code == 200
@@ -179,6 +179,91 @@ def test_relationship_one_filter(client_data):
                     "teacher": None,
                 },
                 {"id": 2, "name": "Physics", "teacherId": 1, "teacher": None},
+            ],
+        },
+    }
+
+
+def test_and(client_data):
+    response = client_data.post(
+        "/graphql",
+        json={
+            "query": """query MyQuery {
+  students(where: {AND: {name: {like: "%ther%"}, id: {gte: 2}}}) {
+    id
+    name
+  }
+}
+                            """,
+        },
+    )
+    assert response.status_code == 200
+    assert response.json() == {
+        "data": {
+            "students": [
+                {
+                    "id": 2,
+                    "name": "Other Name"
+                },
+                {
+                    "id": 3,
+                    "name": "Another Name"
+                }
             ]
         }
     }
+
+
+def test_or(client_data):
+    response = client_data.post(
+        "/graphql",
+        json={
+            "query": """query MyQuery {
+  students(where: {OR: {name: {like: "Another%"}, id: {in: 1}}}) {
+    id
+    name
+  }
+}
+                            """,
+        },
+    )
+    assert response.status_code == 200
+    assert response.json() == {
+        "data": {
+            "students": [
+                {
+                    "id": 1,
+                    "name": "Some Name"
+                },
+                {
+                    "id": 3,
+                    "name": "Another Name"
+                }
+            ]
+        }
+    }
+
+def test_not(client_data):
+    response = client_data.post(
+        "/graphql",
+        json={
+            "query": """query MyQuery {
+  students(where: {NOT: {name: {like: "%ther%"}}}) {
+    id
+    name
+  }
+}
+                            """,
+        },
+    )
+    assert response.status_code == 200
+    assert response.json() == {
+  "data": {
+    "students": [
+      {
+        "id": 1,
+        "name": "Some Name"
+      }
+    ]
+  }
+}
